Api Improvements
Validation and documentation

+++ src/core/BaseController.ts
@@ -14,9 +14,13 @@ export default class BaseController<Model> implements IBaseController {
   getAll = async (req: Request, res: Response) => {
 
     const result: DatabaseResult<Model> = await this.service.getAll()
+
+    if(!result.all){
+      return res.status(500).json({data: 'Error extracting the items'});
+    }
+
     const items = result.all;
 
-    
     res.status(200).json({data: items});
   };
 
@@ -35,11 +39,21 @@ export default class BaseController<Model> implements IBaseController {
 
   delete = async (req: Request, res: Response) => {
     const id = req.params.id
+  
+    const record: DatabaseResult<Model> = await this.service.getById(id)
+    const item: Model = record.one;
+
+    if(!item){
+      return res.status(404).json({data: `Item with id [${id}] not found`});
+    }
 
     const result: DatabaseResult<Model> = await this.service.delete(id)
-    const item: Model = result.one;
+    
+    if(!result || !result.affected){
+      return res.status(500).json({data: `Error deleting item with id [${id}]`});
+    }
 
-    res.status(200).json({data: item});
+    res.status(200).json({data: `Sucessfully deleted item with id [${id}]`});
   };
 
 }
